+++ server/src/controller/commentController.ts
@@ -78,7 +78,6 @@ commentController.post('/', async (c) => {
         }
 
         const payload = c.get('jwtPayload')
-        console.log(payload)
         if (payload)
             body.author = payload.sub
         const comment = await commentRepository.create(body)
@@ -94,6 +93,11 @@ commentController.post('/', async (c) => {
 commentController.put('/:id', async (c) => {
     try {
         const body = await c.req.json<Extract<Comment, typeof comments.$inferSelect>>()
+        const payload = c.get('jwtPayload')
+
+        if (payload.sub !== body.author) {
+            return c.json({ error: 'Brak wymaganych uprawnień do edycji komentarza' }, 401);
+        }
 
         if (!body.id || !body.content) {
             return c.json({ error: 'Brak ID lub treści komentarza' }, 400);
@@ -111,9 +115,16 @@ commentController.put('/:id', async (c) => {
 commentController.delete('/:id', async (c) => {
     try {
         const id = c.req.param('id')
+        const payload = c.get('jwtPayload')
 
         if (isNaN(Number(id))) {
             return c.json({ error: 'Nieprawidłowy identyfikator komentarza' }, 400);
+        }
+
+        const found = await commentRepository.findById(Number(id))
+
+        if (payload.sub !== found?.author.username) {
+            return c.json({ error: 'Brak wymaganych uprawnień do usunięcia komentarza' }, 401);
         }
 
         const comment = await commentRepository.deleteById(Number(id))
